Turned playlist manager into a module (struct with data and function pointers), removing a bunch of exported functions and all those switch statements.
Changed the builtin playlist handler to keep the playlist file open
instead of closing and reopening every track. This reduces the possibility
of errors and makes it more efficient since you don't have to seek to the
correct line each time now. Also we can unlink the randomised version as
soon as we've created it meaning less cleanup.


git-svn-id: http://svn.xiph.org/icecast/trunk@6247 0101bb08-14d6-0310-b084-bc0e0c8e3800