Initial auth merge. Add an auth thread (multiple threads can be done later) which can be used to handle authentication mechanisms without taking locks for long periods.  Non-authenticated mountpoints bypass the auth thread.
The lookup/checking of the source_t is done after the authentication succeeds
so the fallback mechanism does not affect which authenticator is used. This
can be extended to allow us to authenticate in webroot as well. XML re-read
changes will take effect immediately for new listeners but existing listeners
will use the original auth_t (refcounted) when they exit.

htpasswd access has been seperated out from auth.c, and implements an AVL
tree for a faster username lookup.  The htpasswd file timestamp is checked
just in case there are changes made externally


git-svn-id: http://svn.xiph.org/icecast/trunk@9713 0101bb08-14d6-0310-b084-bc0e0c8e3800